(GH-21) Fix: Turning on auth does not allow push
Use verbs to specify that PUT and POST are still allowed, but any other
verbs would fall into the deny rule. As the rules are evaluated in
order, a PUT and POST would be allowed and the deny rule would apply to
all other requests.

Simplify the security aspects by removing other location paths as they
are no longer necessary given this is a solution that works for the
entire site.

Given this change, it means that even the top level url will not be
accessible without credentials, further securing down the site.