(GH-27) Install Chocolatey script
Install Chocolatey with script directly from Chocolatey.Server. Script
will determine if Chocolatey.nupkg is locally cached or not and provide
a proper url for download. This script is meant to be used in the same
way as it is done from https://chocolatey.org/install (the instructions
are the same, the url is different).