(GH-43) Allow Install.ps1 to work w/Basic Auth
When basic authentication is turned on, it is impossible to use the
local install.ps1 script to allow installing Chocolatey internally.
This is due to everything being locked down and requiring a password.

Open things back up a bit to allow folks to be able to reach the
install.ps1 page, download the Chocolatey nupkg, and the 7za.exe file.

Allowing these items through will allow someone to be able to install
Chocolatey from an internal location without a lot of configuration to
set the basic auth in a script or in an infra mgmt tool just to use the
installation script.