(build) Make use of Web Targets NuGet Package
Without this package in place, the build does not work on machines that
are using newer versions of Visual Studio. This is due to the fact that
the projects expect that there are files in certain places, when they
don't actually exist.

Using this NuGet Package means that the required target files are
restored into the packages folder, and can be used from this location,
meaning that things should build correctly regardless of what version
of Visual Studio is installed.